Customers can buy stocks for themselves. || Creates default users on startup.

+++ ds-finance-bank-common/src/main/java/api/BankService.java
@@ -2,6 +2,8 @@ package api;
 
 import jakarta.ejb.Remote;
 
+import java.math.BigDecimal;
+
 public interface BankService {
-    Long getInvestableVolume();
+    BigDecimal getInvestableVolume();
 }
